# Pvwattsv7 speed up shading