--- a/shared/lib_pvshade.cpp
+++ b/shared/lib_pvshade.cpp
@@ -25,7 +25,6 @@ OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <math.h>
 #include <limits>
-#include <sstream>
 #include <vector>
 
 
@@ -277,6 +276,50 @@ void selfshade_xs_horstr(bool landscape,
 		X = (ceil(Hs / L) / (m * r)) * (r - 1.0);
 		S = 1.0 - (floor(g * ndiode / W) / (ndiode * n));
 	}
+}
+
+double ssSkyDiff::lookup(double tilt) {
+    char buf[124];
+    sprintf(buf, "%.3f", tilt);
+    if (SkyDiffTable.find(buf) != SkyDiffTable.end())
+        return SkyDiffTable[buf];
+    return compute(tilt);
+}
+
+double ssSkyDiff::compute(double tilt) {
+    if (gcr == 0)
+        throw std::runtime_error("ssSkyDiff::compute error: gcr required in initialization");
+    // sky diffuse reduction
+    double step = 1.0 / 1000.0;
+    double skydiff = 0.0;
+    double tand_stilt = tand(tilt);
+    double sind_stilt = sind(tilt);
+    double Asky = M_PI + M_PI / pow((1 + pow(tand_stilt, 2)), 0.5);
+    double arg[1000];
+    double gamma[1000];
+    double tan_tilt_gamma[1000];
+    double Asky_shade[1000];
+    for (int n = 0; n < 1000; n++)
+    {
+        arg[n] = (1 / tand_stilt) - (1 / (gcr * sind_stilt * (1 - n * step)));
+        gamma[n] = (-M_PI / 2) + atan(arg[n]);
+        tan_tilt_gamma[n] = tan(tilt * DTOR + gamma[n]);
+        Asky_shade[n] = M_PI + M_PI / pow((1 + tan_tilt_gamma[n] * tan_tilt_gamma[n]), 0.5);
+        if (isnan(Asky_shade[n]))
+        {
+            Asky_shade[n] = Asky;
+        }
+        else if ((tilt * DTOR + gamma[n]) > (M_PI / 2))
+        {
+            Asky_shade[n] = 2 * M_PI - Asky_shade[n];
+        }
+        else {}
+        skydiff += (Asky_shade[n] / Asky) * step;
+    }
+    char buf[124];
+    sprintf(buf, "%.3f", tilt);
+    SkyDiffTable[buf] = skydiff;
+    return skydiff;
 }
 
 // self-shading calculation function