… .vscode/launch.json
… .vscode/tasks.json
… shared/lib_pvshade.h
… ssc/cmod_pvwattsv1.cpp
… ssc/cmod_pvwattsv5.cpp
… ssc/cmod_pvwattsv7.cpp