Pvwattsv7 slowdown from #384 is from the self-shading since turning it off by using fixed roof tracking results in speeds similar to v5.

Refactoring the self-shading code for memoization, I got a speed up for a single pvwattsv7 run with shading from 1.24902 to 0.249561 seconds with fixed tilt. This is still slower than v5 at 0.197782 s. For 1-axis tracking, v5 runs in 0.418802 s  and v7 in 0.458495 s.