@dguittet 2020-06-22
Thanks so much for the thorough and helpful review, Janine. I think it looks in a much cleaner state!

The PV tests are passing on my machine. I'll wait for the Travis CI Check to finish my most recent commit.